Rename index.jsx components to their component name
