+++ src/components/Customer/Review/ShipmentCard/index.jsx
@@ -1,4 +1,4 @@
-import HHGShipmentCard from './HHGShipmentCard';
-import PPMShipmentCard from './PPMShipmentCard';
+import HHGShipmentCard from 'components/Customer/Review/ShipmentCard/HHGShipmentCard/HHGShipmentCard';
+import PPMShipmentCard from 'components/Customer/Review/ShipmentCard/PPMShipmentCard/PPMShipmentCard';
 
 export { HHGShipmentCard, PPMShipmentCard };
